Вынес ViewWithTable в Foundation
