+++ FakeNFT/Scenes /Catalog/NftCollection/NftCollectionView.swift
@@ -55,29 +55,6 @@ final class NftCollectionView: UIView, LoadingView, ViewWithTable {
 
 }
 
-protocol ViewWithTable {
-    var tableView: UITableView { get set}
-    func reload()
-    func update(newIndexes: Range<Int>)
-}
-
-extension ViewWithTable {
-    func reload() {
-        tableView.reloadData()
-    }
-
-    func update(newIndexes: Range<Int>) {
-        if newIndexes.isEmpty { return }
-        tableView.performBatchUpdates {
-            let indexPaths = newIndexes.map { index in
-                IndexPath(row: index, section: 0)
-            }
-            tableView.insertRows(at: indexPaths, with: .automatic)
-        } completion: { _ in
-        }
-    }
-}
-
 extension NftCollectionView: UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         delegate?.numberOfRows ?? 0
